+++ arviz/plots/pairplot.py
@@ -34,8 +34,8 @@ def plot_pair(
     point_estimate=None,
     point_estimate_kwargs=None,
     point_estimate_marker_kwargs=None,
-    true_values=None,
-    true_values_kwargs=None,
+    reference_values=None,
+    reference_values_kwargs=None,
     show=None,
 ):
     """
@@ -101,11 +101,11 @@ def plot_pair(
         Additional keywords passed to ax.vline, ax.hline (matplotlib) or ax.square, Span (bokeh)
     point_estimate_marker_kwargs: dict, optional
         Additional keywords passed to ax.scatter in point estimate plot. Not available in bokeh
-    true_values : dict, optional
-        True values for the plotted variables. The true values will be plotted
+    reference_values : dict, optional
+        Reference values for the plotted variables. The Reference values will be plotted
         using a scatter marker
-    true_values_kwargs : dict, optional
-        Additional keywords passed to ax.scatter or ax.circle in true values plot
+    reference_values_kwargs : dict, optional
+        Additional keywords passed to ax.plot or ax.circle in reference values plot
     show : bool, optional
         Call backend show function.
 
@@ -283,8 +283,8 @@ def plot_pair(
         point_estimate=point_estimate,
         point_estimate_kwargs=point_estimate_kwargs,
         point_estimate_marker_kwargs=point_estimate_marker_kwargs,
-        true_values=true_values,
-        true_values_kwargs=true_values_kwargs,
+        reference_values=reference_values,
+        reference_values_kwargs=reference_values_kwargs,
     )
 
     if backend is None: