+++ arviz/tests/base_tests/test_plots_matplotlib.py
@@ -161,7 +161,6 @@ def test_plot_trace(models, kwargs):
 @pytest.mark.parametrize(
     "combined", [True, False],
 )
-
 def test_plot_trace_legend(compact, combined):
     idata = load_arviz_data("rugby")
     axes = plot_trace(